pyproject.toml
[tool.poetry]
name = "billing-prototype-django"
version = "0.1.0"
description = ""
authors = ["Anton Kovalev <acetonen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-extensions = "*"
django-model-utils = "*"
django-rest-auth = "*"
django-rest-swagger = "*"
django = "==3.1.4"
djangorestframework-jwt = "*"
djangorestframework = "==3.11.1"
drf-yasg = "*"
psycopg2-binary = "*"
django-environ = "*"
django-split-settings = '*'
python-memcached = "*"
django-filter = "*"
pytest-asyncio = "*"
aiohttp = "*"
requests = "*"
typing-extensions = "*"
ipython = "*"
pytest-sugar = "*"
pytest-django = "*"
factory-boy = "==2.12.0"
pytest = "*"
pytest-xdist = "*"
django-debug-toolbar = "==3.1"
pytest-bdd = "*"
uvicorn = "*"
uvloop = "*"
httptools = "*"
pydantic = "^1.7.3"

[tool.poetry.dev-dependencies]
black = "==20.8b1"
mypy = "*"
django-stubs = "*"
djangorestframework-stubs = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
